+++ src/main/java/org/example/dao/ProductMapper.java
@@ -11,7 +11,7 @@ public class ProductMapper implements RowMapper<Product> {
     public Product mapRow(ResultSet resultSet, int i) throws SQLException {
         Product product = new Product();
 
-        product.setId(resultSet.getInt("id"));
+        product.setId(resultSet.getLong("id"));
         product.setName(resultSet.getString("name"));
         product.setDescription(resultSet.getString("description"));
         product.setPrice(resultSet.getInt("price"));
