+++ src/main/java/org/example/models/Person.java
@@ -17,14 +17,17 @@ import javax.validation.constraints.Size;
 public class Person {
 
     @Id
-    private int id;
+    private Long id;
+
     @NotEmpty(message = "Name should not be empty")
     @Size(min = 2, max = 30, message = "Name should be between 2 and 30 characters")
     @Column(name = "name")
     private String name;
+
     @Min(value = 0, message = "Age should be greater than 0")
     @Column(name = "age")
     private int age;
+
     @NotEmpty(message = "Email should not be empty")
     @Email(message = "Email should be valid")
     @Column(name = "email")
